python_backend/requirements.txt
fastapi==0.115.0
passlib==1.7.4
pydantic==2.9.2
SQLAlchemy==2.0.35
uvicorn==0.31.0
jose~=1.0.0
finnhub-python